cleaned up the structure
