+++ .gitignore
@@ -1,4 +1,5 @@
 
-/target
+/target/
 **/*.rs.bk
 Cargo.lock
+/rls/
